Bypass Guard 1.0.9
* Added forced conversion of the country code to uppercase in the native BypassGuard_RequestGeoData(),
since iphub.info began to periodically return the country code in lower case, which breaks the verification logic
* Added cvar 'bypass_guard_check_proxy', which allows you to disable checking for Proxy/VPN. Added for servers from
Russia, because Ukrainian players sometimes cannot access servers in the Russian Federation directly, and use Proxy/VPN for this.
It is not recommended to disable the Proxy/VPN check just like that, this greatly weakens the protection against ban bypass!